+++ lib/server.js
@@ -3,22 +3,26 @@ const httpProxy = require('http-proxy');
 const path = require('path');
 const queryString = require('query-string');
 const fs = require('fs');
-const stream = require('stream');
-const zlib = require('zlib');
+const crypto = require('crypto');
+const url = require('url');
+
 
 const {returnMockData, isTargetReq, recordMockData, updateConfig, decompressData, compressData, streamifyStr} = require('./util');
 const filter = ['xts'];
-let server, proxy;
 
+let server, proxy;
 module.exports = class Server {
 
-    targetUrl = null;
-    requestMatcher = null;
-    overwrite = null;
-    serverPort = null;
-    wsPort = null;
-    mstrMode = null;
-    routes = null;
+    /*
+     *   @param
+     *   targetUrl = null
+     *   requestMatcher = null
+     *   overwrite = null
+     *   serverPort = null
+     *   wsPort = null
+     *   mstrMode = null
+     *   routes = null
+     */
 
     constructor(dir) {
         this.dir = dir;
@@ -33,8 +37,8 @@ module.exports = class Server {
             // use hash of string =  merethod + req.url + reqbody
             // to generate the routes
             // delete property in req body that we don't need;
-            for (prop in filter) {
-                delete body[prop];
+            for(let i = 0; i < filter.length; i++){
+                delete body[filter[i]];
             }
             let reqString = req.method + url.parse(req.url).pathname + JSON.stringify(body);
             let route = crypto.createHash('md5').update(reqString).digest('hex') + ext;
@@ -51,11 +55,16 @@ module.exports = class Server {
 
     parseReqBody(contentType, body) {
         // use indexOf instead of === is because contentType may contain charset parameters.
-        if (contentType.indexOf('application/x-www-form-urlencoded')) {
-            return queryString.parse(body);
-        } else if (contentType.indexOf('application/json')) {
-            return JSON.parse(body);
-        } else {
+        if(!contentType){
+            return {};
+        }
+        if (contentType.indexOf('application/x-www-form-urlencoded') > -1) {
+            body = queryString.parse(body);
+            return body;
+        }else if (contentType.indexOf('application/json') > -1){
+            body = JSON.parse(body);
+            return body;
+        }else {
             return {"ReqBody": body};
         }
     }
@@ -87,9 +96,39 @@ module.exports = class Server {
         Object.assign(this, config);
     }
 
+    returnMockData (route, res) {
+        let dataPath = path.join(this.dir, route);
+        console.log("Return mockData in ", dataPath);
+        let {headers, body} = require(dataPath);;
+        if (headers) {
+            for (let prop in headers) {
+                res.setHeader(prop, headers[prop]);
+            }
+            // set-up the mockId in the response headers
+            let mockId = route.substr(0, route.lastIndexOf('.')) || route;
+            res.setHeader('mockId', mockId);
+        }else{
+            console.log("mockdata header is empty!");
+        }
+        if (body) {
+            let encodingType = headers['content-encoding'];
+            if (typeof body !== 'string') {
+                body = JSON.stringify(body);
+            }
+            body = compressData(encodingType, body);
+            res.setHeader('content-length', body.length);
+            res.write(body);
+            res.end();
+        } else {
+            console.log('mockData body is empty!');
+            res.end('mockData body is empty');
+        }
+    }
+
     start() {
         let self = this;
         proxy = httpProxy.createProxyServer({target: this.targetUrl});
+
         server = http.createServer(function (req, res) {
             let body = [];
             req.on('error', (err) => {
@@ -98,49 +137,47 @@ module.exports = class Server {
                 body.push(chunk);
             }).on('end', () => {
                 body = Buffer.concat(body).toString();
-                let bodyObj = parseReqbody(req['content-type'], body); // parse body as object
-                let route = getRouteByReq(req, bodyObj, self.mstrMode);
+                let bodyObj = self.parseReqBody(req.headers['content-type'], body); // parse body as object
+                let route = self.getRouteByReq(req, bodyObj, self.mstrMode);
                 if (self.routes.includes(route)) {
-                    let dataPath = path.join(self.dir, route);
-                    let data = require(dataPath);
-                    returnMockData(res, data);
+                    self.returnMockData(route, res);
                 } else {
                     let bufferStream = streamifyStr(body);
                     req.bodyObj = bodyObj;
                     proxy.web(req, res, {buffer: bufferStream});
                 }
             });
-
-            proxy.on('proxyRes', function (proxyRes, req, res) {
-                if (isTargetReq(req, self.requestMatcher)) {
-                    let data = {};
-                    let resBody = Buffer.alloc(0);
-                    let route = getRouteByReq(req, req.bodyObj, self.mstrMode);
-                    if (this.needRecord(route)) {
-                        proxyRes.on('data', function (data) {
-                            resBody = Buffer.concat([resBody, data]);
-                        });
-                        proxyRes.on('end', function () {
-                            let encodingType = proxyRes.headers['content-encoding'];
-                            data.headers = proxyRes.headers;
-                            data.body = self.parseResBody(encodingType, resBody);
-                            let storePath = path.join(self.dir, route);
-                            console.log("Record mock data in", storePath);
-                            recordMockData(storePath, data);
-                            updateConfig(self, route);
-                        });
-                    }
-                }
-            });
-
-            proxy.on('error', function (err, req, res) {
-                res.writeHead(500, {
-                    'Content-Type': 'text/plain'
-                });
-                res.end('Both Remote Server and MockData are not available');
-            });
-
         }).listen(self.serverPort);
+
+        proxy.on('proxyRes', function (proxyRes, req, res) {
+            if (isTargetReq(req, self.requestMatcher)) {
+                let data = {};
+                let resBody = Buffer.alloc(0);
+                let route = self.getRouteByReq(req, req.bodyObj, self.mstrMode);
+                if (self.needRecord(route)) {
+                    proxyRes.on('data', function (data) {
+                        resBody = Buffer.concat([resBody, data]);
+                    });
+                    proxyRes.on('end', function () {
+                        let encodingType = proxyRes.headers['content-encoding'];
+                        data.headers = proxyRes.headers;
+                        data.body = self.parseResBody(encodingType, resBody);
+                        let storePath = path.join(self.dir, route);
+                        console.log("Record mock data in", storePath);
+                        recordMockData(storePath, data);
+                        updateConfig(self, route);
+                    });
+                }
+            }
+        });
+
+        proxy.on('error', function (err, req, res) {
+            res.writeHead(500, {
+                'Content-Type': 'text/plain'
+            });
+            res.end('Both Remote Server and MockData are not available');
+        });
+
     }
 
     stop() {
