+++ lib/util.js
@@ -6,7 +6,7 @@ const mkdirp = require('mkdirp');
 const stringifyObject = require('stringify-object');
 const beautify = require('js-beautify').js_beautify;
 const zlib = require('zlib');
-const crypto = require('crypto');
+const stream = require('stream');
 
 module.exports = {
 
@@ -46,32 +46,6 @@ module.exports = {
         }
     },
 
-    returnMockData: function (data, res) {
-        let {headers, body} = data;
-        if (headers) {
-            for (let prop in headers) {
-                res.setHeader(prop, headers[prop]);
-            }
-            // set-up the mockId in the response headers
-            let mockId = route.substr(0, route.lastIndexOf('.')) || route;
-            res.setHeader('mockId', mockId);
-        }else{
-            console.log("mockdata header is empty!");
-        }
-        if (body) {
-            let encodingType = headers['content-encoding'];
-            if (typeof body !== 'string') {
-                body = JSON.stringify(body);
-            }
-            body = module.exports.compressData(encodingType, body);
-            res.setHeader('content-length', body.length);
-            res.write(body);
-            res.end();
-        } else {
-            console.log('mockData body is empty!');
-            res.end();
-        }
-    },
 
     isTargetReq: function (req, requestMatcher) {
         let reqFullUrl = url.parse(req.url).pathname;
@@ -133,7 +107,7 @@ module.exports = {
 
     streamifyStr: function (str) {
         let bufferStream = new stream.PassThrough();
-        bufferStream.end(Buffer.from(body, 'utf-8'));
+        bufferStream.end(Buffer.from(str, 'utf-8'));
         return bufferStream;
     }
 }
